Require to pass an http.Client to all functions
Add a client argument to LookupTerms() and Terms() in order to possibly
use something else then http.DefaultClient.

This is needed in order to test via e.g. go-vcr.

Adjust all tests accordingly and switch to use go-vcr so that if
corresponding fixtures/* does not exist actual HTTP requests are done
against treccani.it. If they exists such cassetes are reused.

In this way we are indipendent on treccani.it for usual testing.

Should fix regressions noticed in the last weeks/months.